chore : @Transactional 추가, 대댓글 삭제 시 조건 추가

+++ src/main/java/com/efub/bageasy/domain/reply/service/ReplyService.java
@@ -9,10 +9,12 @@ import com.efub.bageasy.domain.reply.dto.ReplyRequestDto;
 import com.efub.bageasy.domain.reply.repository.ReplyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class ReplyService {
     private final PostService postService;
@@ -45,6 +47,6 @@ public class ReplyService {
     //대댓글 삭제
     public void deleteReply(Member member, Long replyId) {
         Reply reply = findReply(replyId);
-        replyRepository.delete(reply);
+        if(reply.getMemberId() == member.getMemberId())  replyRepository.delete(reply);
     }
 }
